+++ AdventureBackpacks/AdventureBackpacks.cs
@@ -29,7 +29,7 @@ namespace AdventureBackpacks
         //Module Constants
         private const string _pluginId = "vapok.mods.adventurebackpacks";
         private const string _displayName = "Adventure Backpacks";
-        private const string _version = "1.6.30";
+        private const string _version = "1.7.0";
         
         //Interface Properties
         public string PluginId => _pluginId;